Move network command line tools to a dedicated module
